moved formupload from comp to pages
